requirements.txt
aiomysql==0.0.21
aiosqlite==0.16.1
anyio==3.2.0
argon2-cffi==20.1.0
arrow==1.1.1
asgiref==3.4.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
bleach==3.3.0
bullet==2.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
cryptography==3.4.7
cycler==0.10.0
DateTime==4.3
decorator==5.0.9
defusedxml==0.7.1
dnspython==2.1.0
docx==0.2.4
elasticsearch==7.13.1
email-validator==1.1.3
entrypoints==0.3
fastapi==0.65.2
findspark==1.4.2
Flask==2.0.1
Flask-MySQLdb==0.2.0
h11==0.12.0
httptools==0.2.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
iso8601==0.1.14
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jinja2-time==0.2.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
jwt==1.2.0
kiwisolver==1.3.1
lxml==4.6.3
manage-fastapi==1.0.0
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mysql==0.0.3
mysql-connector-python==8.0.25
mysqlclient==2.0.3
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
opencv-python==4.5.2.54
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
poyo==0.5.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
py4j==0.10.9
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
PyJWT==2.1.0
PyMySQL==0.9.3
pypandoc==1.5
pyparsing==2.4.7
pypika-tortoise==0.1.1
pyrsistent==0.17.3
pyspark==3.1.2
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.18.0
python-slugify==5.0.2
pytz==2021.1
pywin32==301
pywinpty==1.1.2
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.2.1
starlette==0.14.2
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.1.0
tornado==6.1
tortoise-orm==0.17.4
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
websockets==9.1
Werkzeug==2.0.1
zope.interface==5.4.0
